Tepora-app/backend/pyproject.toml
[project]
name = "tepora"
version = "1.0.0"
description = "Local-first personal AI agent with episodic memory"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Tepora Development Team"}
]
keywords = ["ai", "agent", "llm", "local-first", "episodic-memory"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # PyTorch - Core ML framework
    "torch==2.9.1",
    
    # Core ML and NLP
    "transformers==4.57.3",
    
    # LangChain ecosystem
    "langchain-openai==1.1.4",
    "langchain-community==0.4.1",
    "langchain-core==1.2.2",
    "langgraph==1.0.5",
    "langchain-text-splitters==1.1.0",
    "langchain-mcp-adapters==0.2.1",
    "mcp==1.13.1",
    
    # Web framework and utilities
    "fastapi==0.124.4",
    "pydantic==2.12.5",
    "uvicorn[standard]==0.38.0",
    "websockets==15.0.1",
    
    # Vector database
    "chromadb==1.3.7",
    
    # CLI and system utilities
    "psutil==7.1.3",
    
    # HTTP and configuration
    "requests==2.32.5",
    "urllib3==2.3.0",
    "httpx==0.28.1",
    "python-dotenv==1.2.1",
    "PyYAML==6.0.3",
    
    # Web scraping and HTML parsing
    "beautifulsoup4==4.14.3",
    
    # Configuration
    "pydantic-settings>=2.10.1",
    
    # NLP and scientific computing
    "nltk==3.9.2",
    "numpy==2.3.5",
    "scikit-learn==1.6.0",
    "networkx==3.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "ruff==0.14.9",
    "mypy==1.19.1",
]

[project.urls]
Homepage = "https://github.com/tepora-project/tepora"
Documentation = "https://github.com/tepora-project/tepora#readme"
Repository = "https://github.com/tepora-project/tepora"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]  # Line length handled separately

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# ============================================
# UV Configuration
# ============================================
[tool.uv]
# UV設定（Pythonバージョンは requires-python から自動検出）

[tool.uv.sources]
# PyTorch settings removed to use standard PyPI (supports CUDA on Windows/Linux)

